requirements.txt
Django==2.2.3
django-oauth-toolkit==1.2.0
djangorestframework==3.10.0
PyMySQL==0.9.2
django-rest-swagger==2.2.0